Working on the teams app, models are ready now building the admin

+++ news/migrations/0004_auto_20240809_1256.py
@@ -5,12 +5,19 @@ from django.db import migrations
 
 def create_editors_group(apps, schema_editor):
     Group = apps.get_model("auth", "Group")
-    Group.objects.create(name="editors")
+
+    try:
+        Group.objects.create(name="editors")
+
+    except:
+        pass
 
 
 def remove_editors_group(apps, schema_editor):
     Group = apps.get_model("auth", "Group")
-    Group.objects.get(name="editors").delete()
+
+    if Group.objects.get(name="editors").exists():
+        Group.objects.get(name="editors").delete()
 
 
 class Migration(migrations.Migration):
